+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - updated `terminus_cmake_libs.cmake` to look for include paths in `./library` folder.
+- updated `terminus_cmake_libs.cmake` to set the rpath for apple differently than other platforms.
 
 ## [1.0.8] - 2025-11-20